pyproject.toml
[project]
name = "bridge-supa-qgis-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "supabase>=2.27.0",
    "uvicorn>=0.38.0",
]
